+++ etensor/bits.go
@@ -16,7 +16,7 @@ type Bits struct {
 // NewBits returns a new n-dimensional array of bits
 // If strides is nil, row-major strides will be inferred.
 // If names is nil, a slice of empty strings will be created.
-func NewBits(shape, strides []int64, names []string) *Bits {
+func NewBits(shape, strides []int, names []string) *Bits {
 	bt := &Bits{}
 	bt.SetShape(shape, strides, names)
 	ln := bt.Len()
@@ -25,13 +25,13 @@ func NewBits(shape, strides []int64, names []string) *Bits {
 }
 
 // Value returns value at given tensor index
-func (tsr *Bits) Value(i []int64) bool {
+func (tsr *Bits) Value(i []int) bool {
 	j := int(tsr.Offset(i))
 	return tsr.Values.Index(j)
 }
 
 // Set sets value at given tensor index
-func (tsr *Bits) Set(i []int64, val bool) {
+func (tsr *Bits) Set(i []int, val bool) {
 	j := int(tsr.Offset(i))
 	tsr.Values.Set(j, val)
 }
